fix: search templates in the classpath as fallback
When spring boot application is packaged as a jar
application, templates must be stored in the resource
directory since `src/main/webapp` directory will not be
included in the final jar.

By default, we are looking for templates using the default
resource loader which is the application context (therefore,
it depends on the type of ApplicationContext).

To avoid this situation, we now look for templates located
in the classpath as a fallback (i.e when previous
resource loaders fails).